@@ -6,16 +6,12 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
 
+
 class MainController extends AbstractController
 {
-    /**
-     * @Route("/main", name="app_main")
-     */
-    public function index(): Response
+    #[Route('/', name: 'homepage')]
+    public function homePage(): Response
     {
-        return $this->json([
-            'message' => 'Welcome to your new controller!',
-            'path' => 'src/Controller/MainController.php',
-        ]);
+        return $this->render('home.html.twig', []);
     }
 }
